+++ apiEstudo/Domain/Models/IBaseRepository.cs
@@ -6,7 +6,7 @@ namespace apiEstudo.Domain.Models
     public interface IBaseRepository<T>
     {
         public List<T>? GetAll();
-        public T? Get(long id);
-        List<T>? GetListByListId(List<long> listId);
+        public T? Get(int id);
+        List<T>? GetListByListId(List<int> listId);
     }
 }
